feat(RacingGame): 전체 로직 구현 완료
로직 구현 완료
- 기존 변수 및 메서드 활용하여 로직 구성
